REFACTOR Segment handler
